fix #26 : 모달 적용 및 notfound page 적용완료
