+++ src/App.jsx
@@ -7,7 +7,7 @@ import Post from './pages/Post';
 import PostId from './pages/PostId';
 import PostIdEdit from './pages/PostIdEdit';
 import PostIdMessage from './pages/PostIdMessage';
-import NotFoundPage from './pages/NotFoundPage';
+import NotFound from './pages/NotFound';
 
 function App() {
   return (
@@ -24,7 +24,7 @@ function App() {
               <Route path="edit" element={<PostIdEdit />} />
             </Route>
           </Route>
-          <Route path="*" element={<NotFoundPage />} />
+          <Route path="*" element={<NotFound />} />
         </Route>
       </Routes>
     </BrowserRouter>
