+++ src/pages/ListPage.jsx
@@ -44,18 +44,11 @@ function ListPage() {
 
   const getRecentItems = async ({ limit, offset }) => {
     const recipientsAllItems = await getRecipientsAllItems({ limit, offset });
-    // const sortedRecent = JSON.parse(
-    //   JSON.stringify([...recipientsAllItems]),
-    // ).sort((a, b) => new Date(a['createdAt']) - new Date(b['createdAt']));
-    const clonedArray = JSON.parse(JSON.stringify([...recipientsAllItems]));
+    const sortedRecent = JSON.parse(
+      JSON.stringify([...recipientsAllItems]),
+    ).sort((a, b) => new Date(b['createdAt']) - new Date(a['createdAt']));
 
-clonedArray.forEach(item => {
-  console.log(item['createdAt'], new Date(item['createdAt']));
-});
-
-const sortedRecent = clonedArray.sort((a, b) => new Date(a['createdAt']) - new Date(b['createdAt']));
-
-console.log(sortedRecent);
+    console.log(sortedRecent);
 
     if (offset === 0) {
       setRecentItems(sortedRecent);
